Add validation check to submit check service

+++ app/services/return-versions/setup/check/submit-check.service.js
@@ -5,9 +5,13 @@
  * @module SubmitCheckService
  */
 
+const CheckPresenter = require('../../../../presenters/return-versions/setup/check/check.presenter.js')
+const CheckValidation = require('../../../../validators/return-versions/check/check.validator.js')
+const FetchPointsService = require('../fetch-points.service.js')
 const GenerateReturnVersionService = require('./generate-return-version.service.js')
 const PersistReturnVersionService = require('./persist-return-version.service.js')
 const ProcessLicenceReturnLogsService = require('../../../return-logs/process-licence-return-logs.service.js')
+const ReturnRequirementsPresenter = require('../../../../presenters/return-versions/setup/check/returns-requirements.presenter.js')
 const SessionModel = require('../../../../models/session.model.js')
 const VoidReturnLogsService = require('../../../return-logs/void-return-logs.service.js')
 
@@ -27,27 +31,73 @@ const VoidReturnLogsService = require('../../../return-logs/void-return-logs.ser
 async function go(sessionId, userId) {
   const session = await SessionModel.query().findById(sessionId)
 
-  const returnVersionData = await GenerateReturnVersionService.go(session.data, userId)
+  const validationResult = _validate(session)
 
-  await PersistReturnVersionService.go(returnVersionData)
+  if (!validationResult) {
+    const returnVersionData = await GenerateReturnVersionService.go(session.data, userId)
 
-  const oneDayInMilliseconds = 24 * 60 * 60 * 1000
-  const changeDate = new Date(returnVersionData.returnVersion.startDate)
-  changeDate.setTime(changeDate.getTime() - oneDayInMilliseconds)
+    await PersistReturnVersionService.go(returnVersionData)
 
-  if (session.data.journey === 'no-returns-required') {
-    await VoidReturnLogsService.go(
-      session.data.licence.licenceRef,
-      returnVersionData.returnVersion.startDate,
-      returnVersionData.returnVersion.endDate
-    )
+    const oneDayInMilliseconds = 24 * 60 * 60 * 1000
+    const changeDate = new Date(returnVersionData.returnVersion.startDate)
+    changeDate.setTime(changeDate.getTime() - oneDayInMilliseconds)
+
+    if (session.data.journey === 'no-returns-required') {
+      await VoidReturnLogsService.go(
+        session.data.licence.licenceRef,
+        returnVersionData.returnVersion.startDate,
+        returnVersionData.returnVersion.endDate
+      )
+    }
+
+    await ProcessLicenceReturnLogsService.go(returnVersionData.returnVersion.licenceId, changeDate)
+
+    await SessionModel.query().deleteById(sessionId)
+
+    return { licenceId: session.licence.id }
   }
 
-  await ProcessLicenceReturnLogsService.go(returnVersionData.returnVersion.licenceId, changeDate)
+  const returnRequirements = await _returnRequirements(session)
+  const submittedSessionData = CheckPresenter.go(session)
 
-  await SessionModel.query().deleteById(sessionId)
+  return {
+    activeNavBar: 'search',
+    error: validationResult,
+    ...returnRequirements,
+    ...submittedSessionData
+  }
+}
 
-  return session.licence.id
+async function _returnRequirements(session) {
+  const { licence, requirements, journey } = session
+
+  const points = await FetchPointsService.go(licence.id)
+
+  return ReturnRequirementsPresenter.go(requirements, points, journey)
+}
+
+function _validate(session) {
+  const { requirements, quarterlyReturns } = session
+
+  if (!quarterlyReturns) {
+    return null
+  }
+
+  const returnCycles = requirements.map((requirement) => {
+    return requirement.returnsCycle
+  })
+
+  const validation = CheckValidation.go(returnCycles)
+
+  if (!validation.error) {
+    return null
+  }
+
+  const { message } = validation.error.details[0]
+
+  return {
+    text: message
+  }
 }
 
 module.exports = {
